make header Search bar into component
