+++ eq-author/src/App/QuestionnairesPage/QuestionnairesView/Header/index.js
@@ -4,15 +4,10 @@ import styled from "styled-components";
 import { debounce } from "lodash";
 
 import Button from "components/buttons/Button";
-import { Input } from "components/Forms";
-import VisuallyHidden from "components/VisuallyHidden";
+import SearchBar from 'components/SearchBar'
 import AccessFilter from "./AccessFilter";
 
-import { colors } from "constants/theme";
-
 import QuestionnaireSettingsModal from "App/QuestionnaireSettingsModal";
-
-import iconSearch from "./icon-search.svg";
 
 const DEBOUNCE_TIMEOUT = 200;
 
@@ -22,36 +17,6 @@ const Wrapper = styled.div`
   z-index: 1;
   align-items: center;
   justify-content: space-between;
-`;
-
-const Search = styled.div`
-  position: relative;
-  &::before {
-    content: url(${iconSearch});
-    display: inline-block;
-    position: absolute;
-    left: 0.5em;
-    top: 0;
-    bottom: 0;
-    height: 2em;
-    margin: auto;
-  }
-`;
-
-const SearchInput = styled(Input).attrs({
-  type: "search",
-  placeholder: "Search",
-})`
-  width: 20em;
-  padding: 0.6em;
-  line-height: 1;
-  padding-left: 2.5em;
-  border-radius: 4px;
-  border-color: ${colors.bordersLight};
-
-  &:hover {
-    outline: none;
-  }
 `;
 
 const Header = ({
@@ -72,16 +37,8 @@ const Header = ({
   return (
     <>
       <Wrapper>
-        <Search>
-          <VisuallyHidden>
-            <label htmlFor="search">Search</label>
-          </VisuallyHidden>
-          <SearchInput
-            id="search"
-            defaultValue={""}
-            onChange={onSearchChangeDebounced}
-          />
-        </Search>
+
+        <SearchBar onChange={onSearchChangeDebounced}/>
 
         <AccessFilter onToggleFilter={onToggleFilter} isFiltered={isFiltered} />
